agents/visual_search_agent.py
from .base_agent import BaseAgent, AgentContext
from typing import Dict, Any
from langchain import BaseLanguageModel
from PIL import Image
import numpy as np
from io import BytesIO
import requests

class VisualSearchAgent(BaseAgent):
    """Agent specialized in visual search of fashion items."""
    
    def __init__(self, llm: BaseLanguageModel, context: AgentContext):
        super().__init__(llm, context)
        self.vector_store = None  # Will be injected
    
    def process(self, input_data: Dict[str, Any]) -> Dict[str, Any]:
        """Search for similar fashion items based on an image."""
        image_url = input_data.get("image_url")
        if not image_url:
            return {"error": "No image URL provided"}
            
        try:
            # Download and process the image
            response = requests.get(image_url)
            image = Image.open(BytesIO(response.content))
            
            # TODO: Implement actual image processing and vector generation
            # This is a placeholder for the actual implementation
            image_vector = np.random.random(512)  # Replace with real vector
            
            # Search in vector store
            results = self.vector_store.similarity_search(
                query_vector=image_vector,
                k=5  # Return top 5 results
            )
            
            return {"results": results}
            
        except Exception as e:
            return {"error": str(e)}
    
    def get_agent_type(self) -> str:
        return "visual_search"
