requirements.txt
langgraph>=0.1.0
langsmith>=0.1.0
langchain>=0.1.0
streamlit>=1.28.0
elasticsearch>=8.0.0
pinecone-client>=2.2.0
openai>=1.3.0
anthropic>=0.8.0
pydantic>=2.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
